Iris: Move local providers from local_client to local_platform
The local providers (_LocalContainerRuntime, _LocalBundleProvider,
_LocalImageProvider, LocalEnvironmentProvider) are consumed by
LocalVmManager in local_platform.py. Move them there so local_client.py
is just a thin wrapper around ClusterManager.

Also folds make_local_config logic directly into the config helpers,
removing the redundant two-step config creation.
