+++ lib/iris/src/iris/cluster/client/local_client.py
@@ -19,285 +19,41 @@ instead of Docker containers, ensuring local execution follows the same code pat
 as production cluster execution.
 """
 
-import io
-import threading
 import time
-import uuid
-from contextlib import redirect_stderr, redirect_stdout
-from dataclasses import dataclass, field
-from datetime import datetime, timezone
-from pathlib import Path
 from typing import Self
 
 from iris.cluster.client.remote_client import RemoteClusterClient
 from iris.cluster.types import Entrypoint
-from iris.cluster.vm.cluster_manager import ClusterManager, make_local_config
-from iris.cluster.worker.builder import BuildResult
-from iris.cluster.worker.docker import ContainerConfig, ContainerRuntime, ContainerStats, ContainerStatus
-from iris.cluster.worker.worker_types import LogLine
+from iris.cluster.vm.cluster_manager import ClusterManager
 from iris.rpc import cluster_pb2, config_pb2
 from iris.rpc.cluster_connect import ControllerServiceClientSync
 
 
-class _StreamingCapture(io.StringIO):
-    """StringIO subclass that immediately writes lines to container logs."""
-
-    def __init__(self, container: "_LocalContainer", source: str):
-        super().__init__()
-        self._container = container
-        self._source = source
-        self._buffer = ""
-
-    def write(self, s: str) -> int:
-        result = super().write(s)
-        self._buffer += s
-        while "\n" in self._buffer:
-            line, self._buffer = self._buffer.split("\n", 1)
-            self._container._logs.append(
-                LogLine(
-                    timestamp=datetime.now(timezone.utc),
-                    source=self._source,
-                    data=line,
-                )
-            )
-        return result
-
-    def flush_remaining(self):
-        """Flush any remaining buffered content as a final log line."""
-        if self._buffer:
-            self._container._logs.append(
-                LogLine(
-                    timestamp=datetime.now(timezone.utc),
-                    source=self._source,
-                    data=self._buffer,
-                )
-            )
-            self._buffer = ""
-
-
-# =============================================================================
-# Local Providers (private implementation details)
-# =============================================================================
-
-
-class LocalEnvironmentProvider:
-    def __init__(
-        self,
-        cpu: int = 1000,
-        memory_gb: int = 1000,
-        attributes: dict[str, str | int | float] | None = None,
-        device: cluster_pb2.DeviceConfig | None = None,
-    ):
-        self._cpu = cpu
-        self._memory_gb = memory_gb
-        self._attributes = attributes or {}
-        self._device = device
-
-    def probe(self) -> cluster_pb2.WorkerMetadata:
-        if self._device is not None:
-            device = self._device
-        else:
-            device = cluster_pb2.DeviceConfig()
-            device.cpu.CopyFrom(cluster_pb2.CpuDevice(variant="cpu"))
-
-        proto_attrs = {}
-        for key, value in self._attributes.items():
-            if isinstance(value, str):
-                proto_attrs[key] = cluster_pb2.AttributeValue(string_value=value)
-            elif isinstance(value, int):
-                proto_attrs[key] = cluster_pb2.AttributeValue(int_value=value)
-            elif isinstance(value, float):
-                proto_attrs[key] = cluster_pb2.AttributeValue(float_value=value)
-
-        return cluster_pb2.WorkerMetadata(
-            hostname="local",
-            ip_address="127.0.0.1",
-            cpu_count=self._cpu,
-            memory_bytes=self._memory_gb * 1024**3,
-            disk_bytes=100 * 1024**3,  # Default 100GB for local
-            device=device,
-            attributes=proto_attrs,
-        )
-
-
-@dataclass
-class _LocalContainer:
-    config: ContainerConfig
-    _thread: threading.Thread | None = field(default=None, repr=False)
-    _running: bool = False
-    _exit_code: int | None = None
-    _error: str | None = None
-    _logs: list[LogLine] = field(default_factory=list)
-    _killed: threading.Event = field(default_factory=threading.Event)
-
-    def start(self):
-        self._running = True
-        self._thread = threading.Thread(target=self._execute, daemon=True)
-        self._thread.start()
-
-    def _execute(self):
-        from iris.cluster.client.job_info import JobInfo, _parse_ports_from_env, set_job_info
-
-        stdout_capture = _StreamingCapture(self, "stdout")
-        stderr_capture = _StreamingCapture(self, "stderr")
-
-        try:
-            # Build JobInfo from container config env vars
-            env = self.config.env
-            job_info = JobInfo(
-                job_id=env.get("IRIS_JOB_ID", ""),
-                task_id=env.get("IRIS_TASK_ID"),
-                task_index=int(env.get("IRIS_TASK_INDEX", "0")),
-                num_tasks=int(env.get("IRIS_NUM_TASKS", "1")),
-                attempt_id=int(env.get("IRIS_ATTEMPT_ID", "0")),
-                worker_id=env.get("IRIS_WORKER_ID"),
-                controller_address=env.get("IRIS_CONTROLLER_ADDRESS"),
-                ports=_parse_ports_from_env(env),
-            )
-            set_job_info(job_info)
-
-            entrypoint = self.config.entrypoint
-
-            # Check if killed before executing
-            if self._killed.is_set():
-                self._exit_code = 137
-                return
-
-            # Execute based on entrypoint type
-            with redirect_stdout(stdout_capture), redirect_stderr(stderr_capture):
-                if entrypoint.is_callable:
-                    fn, args, kwargs = entrypoint.resolve()
-                    fn(*args, **kwargs)
-                else:
-                    # Command entrypoint: run subprocess with output capture
-                    assert entrypoint.command is not None
-                    import subprocess
-
-                    result = subprocess.run(
-                        entrypoint.command,
-                        capture_output=True,
-                        text=True,
-                        check=False,
-                        env=self.config.env,
-                    )
-                    stdout_capture.write(result.stdout)
-                    stderr_capture.write(result.stderr)
-                    if result.returncode != 0:
-                        raise RuntimeError(f"Command failed with exit code {result.returncode}")
-            self._exit_code = 0
-
-        except Exception as e:
-            self._error = str(e)
-            self._exit_code = 1
-        finally:
-            # Flush any remaining buffered output
-            stdout_capture.flush_remaining()
-            stderr_capture.flush_remaining()
-            self._running = False
-
-    def kill(self):
-        self._killed.set()
-        # Give thread a moment to notice
-        if self._thread and self._thread.is_alive():
-            self._thread.join(timeout=0.5)
-        if self._running:
-            self._running = False
-            self._exit_code = 137
-
-
-class _LocalContainerRuntime(ContainerRuntime):
-    def __init__(self):
-        self._containers: dict[str, _LocalContainer] = {}
-
-    def create_container(self, config: ContainerConfig) -> str:
-        container_id = f"local-{uuid.uuid4().hex[:8]}"
-        self._containers[container_id] = _LocalContainer(config=config)
-        return container_id
-
-    def start_container(self, container_id: str) -> None:
-        self._containers[container_id].start()
-
-    def inspect(self, container_id: str) -> ContainerStatus:
-        c = self._containers.get(container_id)
-        if not c:
-            return ContainerStatus(running=False, exit_code=1, error="container not found")
-        return ContainerStatus(
-            running=c._running,
-            exit_code=c._exit_code,
-            error=c._error,
-        )
-
-    def kill(self, container_id: str, force: bool = False) -> None:
-        del force  # Local containers don't distinguish force vs graceful
-        if container_id in self._containers:
-            self._containers[container_id].kill()
-
-    def remove(self, container_id: str) -> None:
-        self._containers.pop(container_id, None)
-
-    def get_logs(self, container_id: str) -> list[LogLine]:
-        c = self._containers.get(container_id)
-        return c._logs if c else []
-
-    def get_stats(self, container_id: str) -> ContainerStats:
-        del container_id
-        return ContainerStats(memory_mb=100, cpu_percent=10, process_count=1, available=True)
-
-    def list_iris_containers(self, all_states: bool = True) -> list[str]:
-        del all_states
-        return list(self._containers.keys())
-
-    def remove_all_iris_containers(self) -> int:
-        count = len(self._containers)
-        self._containers.clear()
-        return count
-
-
-class _LocalBundleProvider:
-    def __init__(self, bundle_path: Path):
-        self._bundle_path = bundle_path
-
-    def get_bundle(self, gcs_path: str, expected_hash: str | None = None) -> Path:
-        del gcs_path, expected_hash
-        return self._bundle_path
-
-
-class _LocalImageProvider:
-    def build(
-        self,
-        bundle_path: Path,
-        base_image: str,
-        extras: list[str],
-        job_id: str,
-        task_logs=None,
-        pip_packages: list[str] | None = None,
-    ) -> BuildResult:
-        del bundle_path, base_image, extras, job_id, task_logs, pip_packages
-        return BuildResult(
-            image_tag="local:latest",
-            build_time_ms=0,
-            from_cache=True,
-        )
-
-    def protect(self, tag: str) -> None:
-        """No-op for local provider (no eviction)."""
-        del tag
-
-    def unprotect(self, tag: str) -> None:
-        """No-op for local provider (no eviction)."""
-        del tag
-
-
 def _make_local_cluster_config(max_workers: int) -> config_pb2.IrisClusterConfig:
-    """Build a minimal IrisClusterConfig for local execution."""
+    """Build a fully-configured IrisClusterConfig for local execution.
+
+    Sets up controller_vm.local, bundle_prefix, scale groups with local provider,
+    and fast autoscaler evaluation for tests.
+    """
     config = config_pb2.IrisClusterConfig()
+
+    # Configure local controller
+    config.controller_vm.local.port = 0  # auto-assign
+    config.controller_vm.bundle_prefix = ""  # LocalController will set temp path
+
+    # Configure scale group with local provider
     sg = config_pb2.ScaleGroupConfig(
         name="local-cpu",
         min_slices=1,
         max_slices=max_workers,
         accelerator_type=config_pb2.ACCELERATOR_TYPE_CPU,
     )
+    sg.provider.local.SetInParent()
     config.scale_groups["local-cpu"].CopyFrom(sg)
+
+    # Fast autoscaler evaluation for tests
+    config.autoscaler.evaluation_interval_seconds = 0.5
+
     return config
 
 
@@ -328,7 +84,6 @@ class LocalClusterClient:
             A fully initialized LocalClusterClient ready for use
         """
         config = _make_local_cluster_config(max_workers)
-        config = make_local_config(config)
         manager = ClusterManager(config)
         address = manager.start()
         cls._wait_for_worker_registration(address)
